1.添加app管控功能 2.添加动态更新管控列表 3.添加显示被管控应用的桌面小部件 4.添加获取CPU使用率功能 5.添加获取网络速率

+++ app/src/main/java/com/bysj/lizhunan/bean/Used.java
@@ -9,7 +9,7 @@ public class Used {
     private int cpuUsed;
     private int netUsed;
     private int memoryPer;
-    private int cpuPer;
+    private double cpuPer;
     private int netPer;
     private int currMemory;
     private int currCpu;
@@ -47,11 +47,11 @@ public class Used {
         this.memoryPer = memoryPer;
     }
 
-    public int getCpuPer() {
+    public double getCpuPer() {
         return cpuPer;
     }
 
-    public void setCpuPer(int cpuPer) {
+    public void setCpuPer(double cpuPer) {
         this.cpuPer = cpuPer;
     }
